Former-commit-id: 75711b663fce2da01eebc4dc8d9667634a1f2918

+++ app/Http/Controllers/ContractController.php
@@ -7,7 +7,6 @@ use App\Carrier;
 use App\Company;
 use App\Container;
 use App\Contract;
-use App\ContractCarrier;
 use App\ContractLcl;
 use App\Country;
 use App\Currency;
@@ -52,8 +51,7 @@ class ContractController extends Controller
      * @param  Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    function list(Request $request)
-    {
+    function list(Request $request) {
         $results = Contract::filterByCurrentCompany()->filter($request);
 
         return ContractResource::collection($results);
@@ -241,7 +239,7 @@ class ContractController extends Controller
             'carriers' => 'required',
         ]);
 
-        $status = $this->updateStatus($data['expire']);
+        $status = $this->updateStatus($contract, $data);
 
         $contract->update([
             'name' => $data['name'],
@@ -257,18 +255,22 @@ class ContractController extends Controller
         return new ContractResource($contract);
     }
 
-    public function updateStatus($data)
+    public function updateStatus($contract, $data)
     {
 
         $date = date('Y-m-d');
-        $expire = date('Y-m-d', strtotime($data));
-
-        if ($expire <= $date) {
-            $status = 'expired';
-        } else {
-            $status = 'publish';
+        $expire = date('Y-m-d', strtotime($data['expire']));
+        
+        if($contract->status != 'incomplete'){
+            if ($date <= $expire) {
+                $status = 'publish';
+            } else {
+                $status = 'expired';
+            }
+        }else{
+            $status = 'incomplete';
         }
-
+        
         return $status;
     }
 
@@ -368,7 +370,10 @@ class ContractController extends Controller
     public function removefile(Request $request, Contract $contract)
     {
         $media = $contract->getMedia('document')->where('id', $request->input('id'))->first();
-        $media->delete();
+        if(!empty($media) == 0){
+            $media->delete();
+        }
+        
 
         return response()->json(null, 204);
     }
@@ -566,7 +571,7 @@ class ContractController extends Controller
                     'status' => 'incomplete',
                     'type' => $type,
                     'gp_container_id' => 1,
-                    'code' => $request->reference,
+                    'code' => $code,
                     'is_api' => 1,
                 ]);
                 break;
@@ -579,7 +584,8 @@ class ContractController extends Controller
                     'expire' => $request->valid_until,
                     'status' => 'incomplete',
                     'type' => $type,
-                    'code' => $request->reference,
+                    'code' => $code,
+                    'is_api' => 1,
                 ]);
                 break;
         }
@@ -608,7 +614,7 @@ class ContractController extends Controller
                     'user_id' => Auth::user()->id,
                     'created' => date("Y-m-d H:i:s"),
                     'username_load' => 'Not assigned',
-                    'data' => '{"containers": [{"id": 1, "code": "20DV", "name": "20 DV"}, {"id": 2, "code": "40DV", "name": "40 DV"}, {"id": 3, "code": "40HC", "name": "40 HC"}, {"id": 4, "code": "45HC", "name": "45 HC"}, {"id": 5, "code": "40NOR", "name": "40 NOR"}], "group_containers": {"id": 1, "name": "DRY"}}',
+                    'data' => '{"containers": [{"id": 1, "code": "20DV", "name": "20 DV"}, {"id": 2, "code": "40DV", "name": "40 DV"}, {"id": 3, "code": "40HC", "name": "40 HC"}, {"id": 4, "code": "45HC", "name": "45 HC"}, {"id": 5, "code": "40NOR", "name": "40 NOR"}], "group_containers": {"id": 1, "name": "DRY"}, "contract":{"code":'.$contract->code.',"is_api":'.$contract->is_api.'}}',
                     'contract_id' => $contract->id,
                 ]);
                 break;
@@ -664,15 +670,15 @@ class ContractController extends Controller
         $container = Container::get();
 
         $data = $request->validate([
-            'referenceC'       => 'required',
+            'referenceC' => 'required',
             'group_containerC' => 'required',
-            'C20DV'            => 'sometimes|required',
-            'C40DV'            => 'sometimes|required',
-            'C40HC'            => 'sometimes|required',
-            'C40NOR'           => 'sometimes|required',
-            'C45HC'            => 'sometimes|required',
-            'amountC'          => 'sometimes|required',
-            'document'         => 'required',
+            'C20DV' => 'sometimes|required',
+            'C40DV' => 'sometimes|required',
+            'C40HC' => 'sometimes|required',
+            'C40NOR' => 'sometimes|required',
+            'C45HC' => 'sometimes|required',
+            'amountC' => 'sometimes|required',
+            'document' => 'required',
         ]);
 
         $contract->company_user_id = Auth::user()->company_user_id;
@@ -683,6 +689,7 @@ class ContractController extends Controller
         $contract->expire = $validation[1];
         $contract->status = 'publish';
         $contract->gp_container_id = $request->group_containerC;
+        $contract->is_manual = 2;
         $contract->save();
 
         $contract->ContractCarrierSyncSingle($request->carrierR);
@@ -727,7 +734,7 @@ class ContractController extends Controller
         $currencyC = $request->input('currency');
         $amountC = $request->input('amount');
 
-        if (count($calculation_type) > 0) {
+        if (count((array)$calculation_type) > 0) {
             foreach ($calculation_type as $ct => $ctype) {
 
                 if (!empty($request->input('amount'))) {
